refactor(rust): drop the ai-gateway dependency from python-bridge
Move the Responses WebSocket transport (connection type, dial, splice)
from ai-gateway io/responses_ws into core responses::connection, with
send/recv locked on separate halves and a Drop that closes the upstream
socket when the last clone is abandoned.

Move the OCR pipeline (prepare, provider call, document fetch, Azure DI
polling) into core ocr, mirroring the audio_transcription split: core
owns prepare/execute plus the ocr()/ocr_with_observer() entrypoints, the
gateway keeps only its lifecycle hooks (custom loggers/guardrails) and
request type, re-exporting through io/ocr unchanged.

Delete litellm-ai-gateway from python-bridge's dependencies and guard
the direction with manifest tests: python-bridge depends only on core +
python-interop, and core carries no pyo3/pythonize/pyo3-async-runtimes.
