--- litellm-rust/crates/ai-gateway/tests/ocr_lifecycle.rs
@@ -1,928 +1,0 @@
-use litellm_ai_gateway::integrations::types::RequestHooks;
-use litellm_core::request_context::LiteLlmRequestContext;
-use litellm_core::request_context::RequestAttribution;
-use litellm_core::request_options::RequestOptions;
-use std::sync::{Arc, Mutex};
-use std::time::Duration;
-
-use litellm_ai_gateway::integrations::custom_guardrail::{
-    CustomGuardrail, GuardrailContext, GuardrailDecision, GuardrailError, GuardrailEventHook,
-    GuardrailFuture, GuardrailRequest,
-};
-use litellm_ai_gateway::integrations::custom_logger::{
-    CallbackTiming, CallbackValue, CustomLogger, LogFuture, ModelCallDetails,
-};
-use litellm_ai_gateway::ocr::{OcrRequest, ocr, ocr_with_observer};
-use litellm_core::error::Error;
-use litellm_core::provider_callbacks::{
-    CallbackDecision, ProviderAttemptObserver, ProviderError, ProviderPostCall, ProviderPreCall,
-};
-use serde_json::{Map, Value, json};
-use tokio::io::{AsyncReadExt, AsyncWriteExt};
-use tokio::net::{TcpListener, TcpStream};
-
-struct ProviderObserver {
-    events: Arc<Mutex<Vec<&'static str>>>,
-    raw_response: Option<String>,
-    rejected_callback: Option<&'static str>,
-    decision: Option<&'static str>,
-}
-
-impl ProviderAttemptObserver for ProviderObserver {
-    type Error = &'static str;
-
-    async fn pre_call(&mut self, input: &ProviderPreCall) -> Result<CallbackDecision, Self::Error> {
-        assert_eq!(input.model, "mistral-ocr-4-1");
-        assert_eq!(input.call_id, "observer-test");
-        assert_eq!(
-            input.request["document"]["document_url"],
-            "https://example.com/document.pdf"
-        );
-        assert!(input.api_base.ends_with("/v1/ocr"));
-        assert!(
-            input
-                .headers
-                .values()
-                .any(|value| value == "Bearer test-key")
-        );
-        self.events.lock().unwrap().push("pre");
-        match (self.rejected_callback, self.decision) {
-            (Some("pre"), _) => Err("observer failure"),
-            (_, Some("replace_pre")) => Ok(CallbackDecision::Replace {
-                payload: Value::Object(
-                    input
-                        .request
-                        .iter()
-                        .map(|(key, value)| (key.clone(), value.clone()))
-                        .chain(std::iter::once((
-                            "callback_replaced".to_string(),
-                            json!(true),
-                        )))
-                        .collect(),
-                ),
-            }),
-            (_, Some("reject_pre")) => Ok(CallbackDecision::Reject {
-                message: "callback rejected request".to_string(),
-                status_code: Some(400),
-            }),
-            _ => Ok(CallbackDecision::Unchanged),
-        }
-    }
-
-    async fn post_call(
-        &mut self,
-        input: &ProviderPostCall,
-    ) -> Result<CallbackDecision, Self::Error> {
-        self.events.lock().unwrap().push("post");
-        self.raw_response = input.response.as_str().map(str::to_string);
-        match (self.rejected_callback, self.decision) {
-            (Some("post"), _) => Err("observer failure"),
-            (_, Some("replace_post")) => Ok(CallbackDecision::Replace {
-                payload: json!({"pages":[{"index":0,"markdown":"masked"}]}),
-            }),
-            _ => Ok(CallbackDecision::Unchanged),
-        }
-    }
-
-    async fn error(&mut self, input: &ProviderError) -> Result<(), Self::Error> {
-        assert!(input.committed);
-        assert!(!input.message.is_empty());
-        self.events.lock().unwrap().push("error");
-        if self.rejected_callback == Some("error") {
-            Err("observer failure")
-        } else {
-            Ok(())
-        }
-    }
-}
-
-fn observer_request() -> OcrRequest<'static> {
-    OcrRequest {
-        model: "mistral/mistral-ocr-4-1",
-        document: json!({"type":"document_url","document_url":"https://example.com/document.pdf"}),
-        optional_params: Map::new(),
-    }
-}
-
-fn observer_options(api_base: &str) -> RequestOptions {
-    RequestOptions {
-        api_key: Some("test-key".into()),
-        api_base: Some(api_base.into()),
-        custom_llm_provider: Some("mistral".into()),
-        timeout: Some(Duration::from_secs(2)),
-        ..Default::default()
-    }
-}
-
-fn observer_context() -> LiteLlmRequestContext {
-    LiteLlmRequestContext {
-        litellm_call_id: Some("observer-test".into()),
-        ..Default::default()
-    }
-}
-
-async fn observer_case(status: u16, body: &'static str, decision: Option<&'static str>) {
-    let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-    let url = format!("http://{}/v1", listener.local_addr().unwrap());
-    let events = Arc::new(Mutex::new(Vec::new()));
-    let provider_events = Arc::clone(&events);
-    let server = tokio::spawn(async move {
-        let (mut socket, _) = listener.accept().await.unwrap();
-        let request = read_http_request(&mut socket).await;
-        assert!(request.starts_with("POST /v1/ocr "));
-        assert_eq!(
-            request.contains(r#""callback_replaced":true"#),
-            decision == Some("replace_pre")
-        );
-        provider_events.lock().unwrap().push("http");
-        let response = format!(
-            "HTTP/1.1 {status} Test\r\ncontent-length: {}\r\nconnection: close\r\n\r\n{body}",
-            body.len()
-        );
-        socket.write_all(response.as_bytes()).await.unwrap();
-    });
-    let mut observer = ProviderObserver {
-        events: Arc::clone(&events),
-        raw_response: None,
-        rejected_callback: None,
-        decision,
-    };
-    let result = ocr_with_observer(
-        observer_request(),
-        &observer_options(&url),
-        &observer_context(),
-        RequestHooks::default(),
-        &mut observer,
-    )
-    .await;
-    tokio::time::timeout(Duration::from_secs(2), server)
-        .await
-        .unwrap()
-        .unwrap();
-    if status != 200 {
-        assert!(matches!(result, Err(Error::Http { status: actual, .. }) if actual == status));
-        assert_eq!(*events.lock().unwrap(), ["pre", "http", "error"]);
-        assert_eq!(observer.raw_response, None);
-    } else {
-        assert_eq!(*events.lock().unwrap(), ["pre", "http", "post"]);
-        assert_eq!(observer.raw_response.as_deref(), Some(body));
-        if body == "invalid-json" {
-            assert!(matches!(result, Err(Error::InvalidResponse(_))));
-        } else {
-            assert_eq!(
-                result.unwrap()["pages"][0]["markdown"],
-                if decision == Some("replace_post") {
-                    "masked"
-                } else {
-                    "ok"
-                }
-            );
-        }
-    }
-}
-
-#[tokio::test]
-async fn provider_observers_surround_http_and_can_replace_request_or_response() {
-    observer_case(200, r#"{"pages":[{"index":0,"markdown":"ok"}]}"#, None).await;
-    observer_case(200, "invalid-json", None).await;
-    observer_case(401, r#"{"error":"rejected"}"#, None).await;
-    observer_case(
-        200,
-        r#"{"pages":[{"index":0,"markdown":"ok"}]}"#,
-        Some("replace_pre"),
-    )
-    .await;
-    observer_case(
-        200,
-        r#"{"pages":[{"index":0,"markdown":"ok"}]}"#,
-        Some("replace_post"),
-    )
-    .await;
-}
-
-#[tokio::test]
-async fn provider_callback_rejection_stops_before_http() {
-    let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-    let url = format!("http://{}/v1", listener.local_addr().unwrap());
-    let events = Arc::new(Mutex::new(Vec::new()));
-    let mut observer = ProviderObserver {
-        events: Arc::clone(&events),
-        raw_response: None,
-        rejected_callback: None,
-        decision: Some("reject_pre"),
-    };
-
-    let result = ocr_with_observer(
-        observer_request(),
-        &observer_options(&url),
-        &observer_context(),
-        RequestHooks::default(),
-        &mut observer,
-    )
-    .await;
-
-    assert!(
-        matches!(result, Err(Error::InvalidRequest(message)) if message == "callback rejected request")
-    );
-    assert_eq!(*events.lock().unwrap(), ["pre"]);
-    assert!(
-        tokio::time::timeout(Duration::from_millis(50), listener.accept())
-            .await
-            .is_err()
-    );
-}
-
-#[tokio::test]
-async fn invalid_ocr_preparation_does_not_call_observers_or_provider() {
-    let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-    let url = format!("http://{}/v1", listener.local_addr().unwrap());
-    let events = Arc::new(Mutex::new(Vec::new()));
-    let mut observer = ProviderObserver {
-        events: Arc::clone(&events),
-        raw_response: None,
-        rejected_callback: None,
-        decision: None,
-    };
-    let request = OcrRequest {
-        document: json!(42),
-        ..observer_request()
-    };
-    assert!(
-        ocr_with_observer(
-            request,
-            &observer_options(&url),
-            &observer_context(),
-            RequestHooks::default(),
-            &mut observer
-        )
-        .await
-        .is_err()
-    );
-    assert!(events.lock().unwrap().is_empty());
-    assert!(
-        tokio::time::timeout(Duration::from_millis(50), listener.accept())
-            .await
-            .is_err()
-    );
-}
-
-async fn read_http_headers(socket: &mut TcpStream) -> String {
-    let mut request = Vec::new();
-    let mut buffer = [0_u8; 1024];
-    loop {
-        let n = socket.read(&mut buffer).await.expect("reads request");
-        if n == 0 {
-            break;
-        }
-        request.extend_from_slice(&buffer[..n]);
-        if request.windows(4).any(|window| window == b"\r\n\r\n") {
-            break;
-        }
-    }
-    String::from_utf8(request).expect("request is utf8")
-}
-
-async fn read_http_request(socket: &mut TcpStream) -> String {
-    let mut request = Vec::new();
-    let mut buffer = [0_u8; 1024];
-    let header_end = loop {
-        let n = socket.read(&mut buffer).await.expect("reads request");
-        if n == 0 {
-            break request.len();
-        }
-        request.extend_from_slice(&buffer[..n]);
-        if let Some(position) = request.windows(4).position(|window| window == b"\r\n\r\n") {
-            break position + 4;
-        }
-    };
-    let headers = String::from_utf8_lossy(&request[..header_end]);
-    let content_length = headers
-        .lines()
-        .find_map(|line| {
-            let (name, value) = line.split_once(':')?;
-            name.eq_ignore_ascii_case("content-length")
-                .then(|| value.trim().parse::<usize>().ok())
-                .flatten()
-        })
-        .unwrap_or(0);
-    while request.len().saturating_sub(header_end) < content_length {
-        let n = socket.read(&mut buffer).await.expect("reads body");
-        if n == 0 {
-            break;
-        }
-        request.extend_from_slice(&buffer[..n]);
-    }
-    String::from_utf8(request).expect("request is utf8")
-}
-
-#[derive(Clone, Debug, PartialEq)]
-struct RecordedLogEvent {
-    hook: &'static str,
-    model: String,
-    call_type: String,
-    user_id: Option<String>,
-    response_object: Option<String>,
-    error_kind: Option<String>,
-}
-
-#[derive(Default)]
-struct RecordingOcrLogger {
-    events: Mutex<Vec<RecordedLogEvent>>,
-}
-
-impl RecordingOcrLogger {
-    fn events(&self) -> Vec<RecordedLogEvent> {
-        self.events.lock().unwrap().clone()
-    }
-}
-
-impl CustomLogger for RecordingOcrLogger {
-    fn async_log_success_event<'a>(
-        &'a self,
-        model_call_details: &'a ModelCallDetails,
-        response_obj: &'a CallbackValue,
-        _timing: CallbackTiming,
-    ) -> LogFuture<'a> {
-        Box::pin(async move {
-            self.events.lock().unwrap().push(RecordedLogEvent {
-                hook: "async_log_success_event",
-                model: model_call_details.model.clone(),
-                call_type: model_call_details.call_type.to_string(),
-                user_id: model_call_details.metadata.user_api_key_user_id.clone(),
-                response_object: Some(response_obj.object.clone()),
-                error_kind: None,
-            });
-            Ok(())
-        })
-    }
-
-    fn async_log_failure_event<'a>(
-        &'a self,
-        model_call_details: &'a ModelCallDetails,
-        response_obj: Option<&'a CallbackValue>,
-        _timing: CallbackTiming,
-    ) -> LogFuture<'a> {
-        Box::pin(async move {
-            self.events.lock().unwrap().push(RecordedLogEvent {
-                hook: "async_log_failure_event",
-                model: model_call_details.model.clone(),
-                call_type: model_call_details.call_type.to_string(),
-                user_id: model_call_details.metadata.user_api_key_user_id.clone(),
-                response_object: response_obj.map(|value| value.object.clone()),
-                error_kind: model_call_details
-                    .failure_error
-                    .as_ref()
-                    .map(|error| error.kind.clone()),
-            });
-            Ok(())
-        })
-    }
-}
-
-struct RecordingOcrGuardrail {
-    hooks: Vec<GuardrailEventHook>,
-    events: Mutex<Vec<&'static str>>,
-    block_pre_call: bool,
-    block_during_call: bool,
-}
-
-impl RecordingOcrGuardrail {
-    fn new(hooks: Vec<GuardrailEventHook>) -> Self {
-        Self {
-            hooks,
-            events: Mutex::new(Vec::new()),
-            block_pre_call: false,
-            block_during_call: false,
-        }
-    }
-
-    fn blocking_pre_call() -> Self {
-        Self {
-            hooks: vec![GuardrailEventHook::PreCall],
-            events: Mutex::new(Vec::new()),
-            block_pre_call: true,
-            block_during_call: false,
-        }
-    }
-
-    fn blocking_during_call() -> Self {
-        Self {
-            hooks: vec![GuardrailEventHook::DuringCall],
-            events: Mutex::new(Vec::new()),
-            block_pre_call: false,
-            block_during_call: true,
-        }
-    }
-
-    fn events(&self) -> Vec<&'static str> {
-        self.events.lock().unwrap().clone()
-    }
-}
-
-impl CustomGuardrail for RecordingOcrGuardrail {
-    fn guardrail_name(&self) -> &str {
-        "recording-ocr-guardrail"
-    }
-
-    fn supported_event_hooks(&self) -> &[GuardrailEventHook] {
-        &self.hooks
-    }
-
-    fn async_pre_call_hook<'a>(
-        &'a self,
-        _context: &'a GuardrailContext,
-        mut request: GuardrailRequest,
-    ) -> GuardrailFuture<'a> {
-        Box::pin(async move {
-            self.events.lock().unwrap().push("async_pre_call_hook");
-            if self.block_pre_call {
-                return Ok(GuardrailDecision::Block(GuardrailError::blocked(
-                    "blocked before provider",
-                )));
-            }
-            request.data["document"]["guarded_pre"] = json!(true);
-            Ok(GuardrailDecision::Mask(request))
-        })
-    }
-
-    fn async_moderation_hook<'a>(
-        &'a self,
-        _context: &'a GuardrailContext,
-        mut request: GuardrailRequest,
-    ) -> GuardrailFuture<'a> {
-        Box::pin(async move {
-            self.events.lock().unwrap().push("async_moderation_hook");
-            if self.block_during_call {
-                return Ok(GuardrailDecision::Block(GuardrailError::blocked(
-                    "blocked before provider",
-                )));
-            }
-            request.data["body"]["guarded_during"] = json!(true);
-            Ok(GuardrailDecision::Mask(request))
-        })
-    }
-}
-
-fn base_ocr_request(model: &str) -> (OcrRequest<'_>, RequestOptions) {
-    (
-        OcrRequest {
-            model,
-            document: json!({
-                "type": "document_url",
-                "document_url": "https://example.com/doc.pdf"
-            }),
-            optional_params: Map::new(),
-        },
-        RequestOptions {
-            api_key: Some("sk-test".to_string()),
-            ..Default::default()
-        },
-    )
-}
-
-#[tokio::test]
-async fn reducto_during_call_guardrail_blocks_before_upload() {
-    let listener = TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("test listener binds");
-    let address = listener.local_addr().expect("listener has local address");
-    let api_base = format!("http://{address}");
-    let guardrail = Arc::new(RecordingOcrGuardrail::blocking_during_call());
-    let (mut request, mut options) = base_ocr_request("reducto/parse-v3");
-    options.api_base = Some(&api_base).map(|value| value.to_string());
-    request.document = json!({
-        "type": "document_url",
-        "document_url": "data:application/pdf;base64,JVBERi0xLjQ="
-    });
-    let hooks = RequestHooks {
-        guardrails: vec![guardrail.clone()],
-        ..Default::default()
-    };
-
-    let error = ocr(
-        request,
-        &options,
-        &LiteLlmRequestContext {
-            ..Default::default()
-        },
-        hooks,
-    )
-    .await
-    .expect_err("guardrail blocks upload");
-
-    assert!(matches!(error, Error::InvalidRequest(_)));
-    assert_eq!(guardrail.events(), vec!["async_moderation_hook"]);
-    let accepted = tokio::time::timeout(Duration::from_millis(100), listener.accept()).await;
-    assert!(accepted.is_err(), "upload socket should not be touched");
-}
-
-#[tokio::test]
-async fn reducto_upload_error_body_is_truncated() {
-    let listener = TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("test listener binds");
-    let address = listener.local_addr().expect("listener has local address");
-    let server = tokio::spawn(async move {
-        let (mut socket, _) = listener.accept().await.expect("accepts upload request");
-        let _request = read_http_request(&mut socket).await;
-        let body = "x".repeat(300);
-        let response = format!(
-            "HTTP/1.1 500 Internal Server Error\r\ncontent-length: {}\r\nconnection: close\r\n\r\n{}",
-            body.len(),
-            body
-        );
-        socket
-            .write_all(response.as_bytes())
-            .await
-            .expect("writes upload response");
-    });
-    let api_base = format!("http://{address}");
-    let (mut request, mut options) = base_ocr_request("reducto/parse-v3");
-    options.api_base = Some(&api_base).map(|value| value.to_string());
-    request.document = json!({
-        "type": "document_url",
-        "document_url": "data:application/pdf;base64,JVBERi0xLjQ="
-    });
-
-    let error = ocr(
-        request,
-        &options,
-        &LiteLlmRequestContext {
-            ..Default::default()
-        },
-        RequestHooks::default(),
-    )
-    .await
-    .expect_err("upload should fail");
-
-    assert!(
-        matches!(error, Error::Http { status: 500, body } if body.chars().count() < 300 && body.ends_with("... (truncated)"))
-    );
-    server.await.expect("server task completes");
-}
-
-#[tokio::test]
-async fn ocr_lifecycle_runs_pre_during_and_success_hooks() {
-    let listener = TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("test listener binds");
-    let addr = listener.local_addr().expect("listener has local addr");
-
-    let server = tokio::spawn(async move {
-        let (mut socket, _) = listener.accept().await.expect("accepts one request");
-        let request = read_http_request(&mut socket).await;
-        let response_body = r#"{"pages":[{"index":0,"markdown":"ok"}],"model":"mistral-ocr-latest","usage_info":{"pages_processed":1}}"#;
-        let response = format!(
-            "HTTP/1.1 200 OK\r\ncontent-type: application/json\r\ncontent-length: {}\r\nconnection: close\r\n\r\n{}",
-            response_body.len(),
-            response_body
-        );
-        socket
-            .write_all(response.as_bytes())
-            .await
-            .expect("writes response");
-        request
-    });
-
-    let logger = Arc::new(RecordingOcrLogger::default());
-    let guardrail = Arc::new(RecordingOcrGuardrail::new(vec![
-        GuardrailEventHook::PreCall,
-        GuardrailEventHook::DuringCall,
-    ]));
-    let response = ocr(
-        OcrRequest {
-            model: "mistral-ocr-latest",
-            document: json!({
-                "type": "document_url",
-                "document_url": "https://example.com/doc.pdf"
-            }),
-            optional_params: Map::new(),
-        },
-        &RequestOptions {
-            api_key: (Some("sk-test")).map(|value| value.to_string()),
-            api_base: (Some(&format!("http://{addr}"))).map(|value| value.to_string()),
-            custom_llm_provider: (Some("mistral")).map(|value| value.to_string()),
-            extra_headers: None,
-            timeout: Some(Duration::from_secs(5)),
-            ..Default::default()
-        },
-        &LiteLlmRequestContext {
-            attribution: RequestAttribution {
-                user_api_key_user_id: Some("user-1".to_string()),
-                ..Default::default()
-            },
-            litellm_call_id: (Some("ocr-call-1")).map(|value| value.to_string()),
-            ..Default::default()
-        },
-        RequestHooks {
-            callbacks: vec![logger.clone()],
-            guardrails: vec![guardrail.clone()],
-        },
-    )
-    .await
-    .expect("ocr request succeeds");
-
-    assert_eq!(response["pages"][0]["markdown"], "ok");
-    assert_eq!(
-        guardrail.events(),
-        vec!["async_pre_call_hook", "async_moderation_hook"]
-    );
-    assert_eq!(
-        logger.events(),
-        vec![RecordedLogEvent {
-            hook: "async_log_success_event",
-            model: "mistral-ocr-latest".to_string(),
-            call_type: "ocr".to_string(),
-            user_id: Some("user-1".to_string()),
-            response_object: Some("ocr".to_string()),
-            error_kind: None,
-        }]
-    );
-
-    let request = server.await.expect("server task completes");
-    assert!(request.contains(r#""guarded_pre":true"#), "{request}");
-    assert!(request.contains(r#""guarded_during":true"#), "{request}");
-}
-
-#[tokio::test]
-async fn ocr_lifecycle_runs_failure_hook_on_provider_error() {
-    let listener = TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("test listener binds");
-    let addr = listener.local_addr().expect("listener has local addr");
-
-    let server = tokio::spawn(async move {
-        let (mut socket, _) = listener.accept().await.expect("accepts one request");
-        let _request = read_http_request(&mut socket).await;
-        let response_body = "provider failed";
-        let response = format!(
-            "HTTP/1.1 500 Internal Server Error\r\ncontent-length: {}\r\nconnection: close\r\n\r\n{}",
-            response_body.len(),
-            response_body
-        );
-        socket
-            .write_all(response.as_bytes())
-            .await
-            .expect("writes response");
-    });
-
-    let logger = Arc::new(RecordingOcrLogger::default());
-    let err = ocr(
-        OcrRequest {
-            model: "mistral-ocr-latest",
-            document: json!({
-                "type": "document_url",
-                "document_url": "https://example.com/doc.pdf"
-            }),
-            optional_params: Map::new(),
-        },
-        &RequestOptions {
-            api_key: (Some("sk-test")).map(|value| value.to_string()),
-            api_base: (Some(&format!("http://{addr}"))).map(|value| value.to_string()),
-            custom_llm_provider: (Some("mistral")).map(|value| value.to_string()),
-            extra_headers: None,
-            timeout: Some(Duration::from_secs(5)),
-            ..Default::default()
-        },
-        &LiteLlmRequestContext {
-            attribution: RequestAttribution::default(),
-            litellm_call_id: (Some("ocr-call-2")).map(|value| value.to_string()),
-            ..Default::default()
-        },
-        RequestHooks {
-            callbacks: vec![logger.clone()],
-            guardrails: Vec::new(),
-        },
-    )
-    .await
-    .expect_err("provider error propagates");
-
-    assert!(matches!(err, Error::Http { status: 500, .. }));
-    server.await.expect("server task completes");
-    assert_eq!(
-        logger.events(),
-        vec![RecordedLogEvent {
-            hook: "async_log_failure_event",
-            model: "mistral-ocr-latest".to_string(),
-            call_type: "ocr".to_string(),
-            user_id: None,
-            response_object: Some("error".to_string()),
-            error_kind: Some("HttpError".to_string()),
-        }]
-    );
-}
-
-#[tokio::test]
-async fn ocr_lifecycle_pre_call_block_skips_provider_socket() {
-    let listener = TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("test listener binds");
-    let addr = listener.local_addr().expect("listener has local addr");
-    let logger = Arc::new(RecordingOcrLogger::default());
-    let guardrail = Arc::new(RecordingOcrGuardrail::blocking_pre_call());
-
-    let err = ocr(
-        OcrRequest {
-            model: "mistral-ocr-latest",
-            document: json!({
-                "type": "document_url",
-                "document_url": "https://example.com/doc.pdf"
-            }),
-            optional_params: Map::new(),
-        },
-        &RequestOptions {
-            api_key: (Some("sk-test")).map(|value| value.to_string()),
-            api_base: (Some(&format!("http://{addr}"))).map(|value| value.to_string()),
-            custom_llm_provider: (Some("mistral")).map(|value| value.to_string()),
-            extra_headers: None,
-            timeout: Some(Duration::from_millis(100)),
-            ..Default::default()
-        },
-        &LiteLlmRequestContext {
-            attribution: RequestAttribution::default(),
-            litellm_call_id: (Some("ocr-call-3")).map(|value| value.to_string()),
-            ..Default::default()
-        },
-        RequestHooks {
-            callbacks: vec![logger.clone()],
-            guardrails: vec![guardrail.clone()],
-        },
-    )
-    .await
-    .expect_err("guardrail blocks request");
-
-    assert!(matches!(err, Error::InvalidRequest(_)));
-    assert_eq!(guardrail.events(), vec!["async_pre_call_hook"]);
-    assert_eq!(
-        logger.events(),
-        vec![RecordedLogEvent {
-            hook: "async_log_failure_event",
-            model: "mistral-ocr-latest".to_string(),
-            call_type: "ocr".to_string(),
-            user_id: None,
-            response_object: Some("error".to_string()),
-            error_kind: Some("InvalidRequest".to_string()),
-        }]
-    );
-    let accepted = tokio::time::timeout(Duration::from_millis(100), listener.accept()).await;
-    assert!(accepted.is_err(), "provider socket should not be touched");
-}
-
-#[tokio::test]
-async fn ocr_does_not_duplicate_authorization_header_when_header_is_supplied() {
-    let listener = TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("test listener binds");
-    let addr = listener.local_addr().expect("listener has local addr");
-
-    let server = tokio::spawn(async move {
-        let (mut socket, _) = listener.accept().await.expect("accepts one request");
-        let request = read_http_headers(&mut socket).await;
-        let response_body = r#"{"pages":[{"index":0,"markdown":"ok"}],"model":"mistral-ocr-latest","usage_info":{"pages_processed":1}}"#;
-        let response = format!(
-            "HTTP/1.1 200 OK\r\ncontent-type: application/json\r\ncontent-length: {}\r\nconnection: close\r\n\r\n{}",
-            response_body.len(),
-            response_body
-        );
-        socket
-            .write_all(response.as_bytes())
-            .await
-            .expect("writes response");
-        request
-    });
-
-    let mut headers = Map::new();
-    headers.insert(
-        "Authorization".to_string(),
-        Value::String("Bearer sk-from-python".to_string()),
-    );
-    headers.insert(
-        "x-trace-id".to_string(),
-        Value::String("trace-1".to_string()),
-    );
-
-    let response = ocr(
-        OcrRequest {
-            model: "mistral-ocr-latest",
-            document: json!({
-                "type": "document_url",
-                "document_url": "https://example.com/doc.pdf"
-            }),
-            optional_params: Map::new(),
-        },
-        &RequestOptions {
-            api_key: (Some("sk-for-rust-fallback")).map(|value| value.to_string()),
-            api_base: (Some(&format!("http://{addr}"))).map(|value| value.to_string()),
-            custom_llm_provider: (Some("mistral")).map(|value| value.to_string()),
-            extra_headers: Some(headers),
-            timeout: Some(Duration::from_secs(5)),
-            ..Default::default()
-        },
-        &LiteLlmRequestContext {
-            attribution: RequestAttribution::default(),
-            litellm_call_id: None,
-            ..Default::default()
-        },
-        RequestHooks {
-            callbacks: Vec::new(),
-            guardrails: Vec::new(),
-        },
-    )
-    .await
-    .expect("ocr request succeeds");
-
-    assert_eq!(response["pages"][0]["markdown"], "ok");
-
-    let request = server.await.expect("server task completes");
-    let authorization_count = request
-        .lines()
-        .filter(|line| line.to_ascii_lowercase().starts_with("authorization:"))
-        .count();
-    assert_eq!(authorization_count, 1, "{request}");
-    assert!(
-        request.contains("authorization: Bearer sk-from-python")
-            || request.contains("Authorization: Bearer sk-from-python"),
-        "{request}"
-    );
-}
-
-#[tokio::test]
-async fn document_intelligence_poll_uses_resolved_subscription_key() {
-    let listener = TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("test listener binds");
-    let addr = listener.local_addr().expect("listener has local addr");
-    let operation_url = format!("http://{addr}/operations/1");
-
-    let server = tokio::spawn(async move {
-        let (mut post_socket, _) = listener.accept().await.expect("accepts post request");
-        let post_request = read_http_headers(&mut post_socket).await;
-        let post_response = format!(
-            "HTTP/1.1 202 Accepted\r\noperation-location: {operation_url}\r\ncontent-length: 0\r\nconnection: close\r\n\r\n"
-        );
-        post_socket
-            .write_all(post_response.as_bytes())
-            .await
-            .expect("writes post response");
-
-        let (mut poll_socket, _) = listener.accept().await.expect("accepts poll request");
-        let poll_request = read_http_headers(&mut poll_socket).await;
-        let response_body = r#"{"status":"succeeded","analyzeResult":{"pages":[{"pageNumber":1,"lines":[{"content":"ok"}]}]}}"#;
-        let poll_response = format!(
-            "HTTP/1.1 200 OK\r\ncontent-type: application/json\r\ncontent-length: {}\r\nconnection: close\r\n\r\n{}",
-            response_body.len(),
-            response_body
-        );
-        poll_socket
-            .write_all(poll_response.as_bytes())
-            .await
-            .expect("writes poll response");
-        (post_request, poll_request)
-    });
-
-    let response = ocr(
-        OcrRequest {
-            model: "doc-intelligence/prebuilt-read",
-            document: json!({
-                "type": "document_url",
-                "document_url": "https://example.com/doc.pdf"
-            }),
-            optional_params: Map::new(),
-        },
-        &RequestOptions {
-            api_key: (Some("di-key")).map(|value| value.to_string()),
-            api_base: (Some(&format!("http://{addr}"))).map(|value| value.to_string()),
-            custom_llm_provider: (Some("azure_ai")).map(|value| value.to_string()),
-            extra_headers: None,
-            timeout: Some(Duration::from_secs(5)),
-            ..Default::default()
-        },
-        &LiteLlmRequestContext {
-            attribution: RequestAttribution::default(),
-            litellm_call_id: None,
-            ..Default::default()
-        },
-        RequestHooks {
-            callbacks: Vec::new(),
-            guardrails: Vec::new(),
-        },
-    )
-    .await
-    .expect("document intelligence request succeeds");
-
-    assert_eq!(response["pages"][0]["markdown"], "ok");
-
-    let (post_request, poll_request) = server.await.expect("server task completes");
-    assert!(
-        post_request
-            .to_ascii_lowercase()
-            .contains("ocp-apim-subscription-key: di-key"),
-        "{post_request}"
-    );
-    assert!(
-        poll_request
-            .to_ascii_lowercase()
-            .contains("ocp-apim-subscription-key: di-key"),
-        "{poll_request}"
-    );
-}
