+++ litellm-rust/crates/core/src/constants.rs
@@ -32,6 +32,18 @@ pub(crate) const CHAT_COMPLETIONS_CONNECT_TIMEOUT_SECS: u64 = 10;
 
 pub(crate) const AUDIO_TRANSCRIPTION_TIMEOUT_SECS: u64 = 600;
 
+/// Full-request timeout ceiling for OCR provider calls, in seconds. The
+/// per-request timeout from the caller still overrides this on the request
+/// builder.
+pub(crate) const OCR_TIMEOUT_SECS: u64 = 600;
+
+/// Connect timeout for the Responses WebSocket upstream dial, in seconds.
+pub(crate) const RESPONSES_WS_CONNECT_TIMEOUT_SECS: u64 = 10;
+
+/// Idle timeout that ends a Responses WebSocket splice when neither side
+/// produces an event, in seconds.
+pub(crate) const RESPONSES_WS_IDLE_TIMEOUT_SECS: u64 = 300;
+
 /// `object` field every non-streaming chat completion response carries.
 pub const CHAT_COMPLETION_OBJECT: &str = "chat.completion";
