+++ litellm-rust/crates/core/src/responses/mod.rs
@@ -1,3 +1,4 @@
+pub mod connection;
 pub mod instrumentation;
 pub mod types;
 pub mod websocket;
