+++ litellm-rust/crates/python-bridge/Cargo.toml
@@ -14,17 +14,12 @@ default = ["abi3"]
 abi3 = ["pyo3/abi3-py310"]
 extension-module = ["pyo3/extension-module"]
 panic-test = []
-trace-parity = [
-    "dep:tracing",
-    "litellm-core/observability",
-    "litellm-ai-gateway/trace-parity",
-]
+trace-parity = ["litellm-core/observability"]
 
 [dependencies]
 futures-util.workspace = true
-tracing = { workspace = true, optional = true }
+tracing.workspace = true
 litellm-core = { workspace = true, features = ["bedrock-auth"] }
-litellm-ai-gateway = { workspace = true, default-features = false }
 litellm-python-interop.workspace = true
 pyo3.workspace = true
 pyo3-async-runtimes.workspace = true
